merge v6.0 to release/v6.0

+++ src/Raven.Server/Smuggler/Documents/Actions/DatabaseRecordActions.cs
@@ -58,6 +58,7 @@ public sealed class DatabaseRecordActions : IDatabaseRecordActions
             return;
 
         var configuration = DatabasesLandlord.CreateDatabaseConfiguration(_server, _name, databaseRecord.Settings);
+        var authenticationEnabled = _server.Configuration.Security.AuthenticationEnabled;
 
         if (databaseRecord.ConflictSolverConfig != null && databaseRecordItemType.HasFlag(DatabaseRecordItemType.ConflictSolverConfig))
         {
@@ -90,25 +91,32 @@ public sealed class DatabaseRecordActions : IDatabaseRecordActions
 
         if (databaseRecord.PeriodicBackups.Count > 0 && databaseRecordItemType.HasFlag(DatabaseRecordItemType.PeriodicBackups))
         {
-            if (_log.IsInfoEnabled)
-                _log.Info("Configuring periodic backups configuration from smuggler");
-
-            foreach (var backupConfig in databaseRecord.PeriodicBackups)
+            if (authenticationEnabled && CanAccess(authorizationStatus) == false)
             {
-                _currentDatabaseRecord?.PeriodicBackups.ForEach(x =>
-                {
-                    if (x.Name.Equals(backupConfig.Name, StringComparison.OrdinalIgnoreCase))
-                    {
-                        tasks.Add(_server.SendToLeaderAsync(new DeleteOngoingTaskCommand(x.TaskId, OngoingTaskType.Backup, _name, RaftIdGenerator.DontCareId)));
-                    }
-                });
-
-                backupConfig.TaskId = 0;
-                backupConfig.Disabled = true;
-                tasks.Add(_server.SendToLeaderAsync(new UpdatePeriodicBackupCommand(backupConfig, _name, RaftIdGenerator.DontCareId)));
+                result.AddError("Import of periodic backup was skipped due to insufficient permissions on your current certificate.");
+                result.DatabaseRecord.ErroredCount++;
             }
+            else
+            {
+                if (_log.IsInfoEnabled)
+                    _log.Info("Configuring periodic backups configuration from smuggler");
 
-            result.DatabaseRecord.PeriodicBackupsUpdated = true;
+                foreach (var backupConfig in databaseRecord.PeriodicBackups)
+                {
+                    _currentDatabaseRecord?.PeriodicBackups.ForEach(x =>
+                    {
+                        if (x.Name.Equals(backupConfig.Name, StringComparison.OrdinalIgnoreCase))
+                        {
+                            tasks.Add(_server.SendToLeaderAsync(new DeleteOngoingTaskCommand(x.TaskId, OngoingTaskType.Backup, _name, RaftIdGenerator.DontCareId)));
+                        }
+                    });
+
+                    backupConfig.TaskId = 0;
+                    backupConfig.Disabled = true;
+                    tasks.Add(_server.SendToLeaderAsync(new UpdatePeriodicBackupCommand(backupConfig, _name, RaftIdGenerator.DontCareId)));
+                }
+                result.DatabaseRecord.PeriodicBackupsUpdated = true;
+            }
         }
 
         if (databaseRecord.SinkPullReplications.Count > 0 && databaseRecordItemType.HasFlag(DatabaseRecordItemType.SinkPullReplications))
@@ -313,28 +321,36 @@ public sealed class DatabaseRecordActions : IDatabaseRecordActions
 
         if (databaseRecord.Revisions != null && databaseRecordItemType.HasFlag(DatabaseRecordItemType.Revisions))
         {
-            if (_log.IsInfoEnabled)
-                _log.Info("Configuring revisions from smuggler");
-
-            if (_currentDatabaseRecord?.Revisions != null)
+            if (authenticationEnabled && CanAccess(authorizationStatus) == false)
             {
-                foreach (var collection in _currentDatabaseRecord.Revisions.Collections)
+                result.AddError("Import of Revision configuration was skipped due to insufficient permissions on your current certificate.");
+                result.DatabaseRecord.ErroredCount++;
+            }
+            else
+            {
+                if (_log.IsInfoEnabled)
+                    _log.Info("Configuring revisions from smuggler");
+
+                if (_currentDatabaseRecord?.Revisions != null)
                 {
-                    if (databaseRecord.Revisions.Collections.TryGetValue(collection.Key, out var collectionConfiguration) == false)
+                    foreach (var collection in _currentDatabaseRecord.Revisions.Collections)
                     {
-                        databaseRecord.Revisions.Collections.Add(collection.Key, collection.Value);
-                    }
-                    else
-                    {
-                        if (collectionConfiguration.Equals(collection.Value) == false)
-                            result.AddWarning($"Revisions configuration of collection '{collection.Key}' already exist on the destination Database Record. " +
-                                              "Configuring this revisions from smuggler was skipped, even though the configuration differed from the configuration in the target database record");
+                        if (databaseRecord.Revisions.Collections.TryGetValue(collection.Key, out var collectionConfiguration) == false)
+                        {
+                            databaseRecord.Revisions.Collections.Add(collection.Key, collection.Value);
+                        }
+                        else
+                        {
+                            if (collectionConfiguration.Equals(collection.Value) == false)
+                                result.AddWarning($"Revisions configuration of collection '{collection.Key}' already exist on the destination Database Record. " +
+                                                  "Configuring this revisions from smuggler was skipped, even though the configuration differed from the configuration in the target database record");
+                        }
                     }
                 }
-            }
 
-            tasks.Add(_server.SendToLeaderAsync(new EditRevisionsConfigurationCommand(databaseRecord.Revisions, _name, RaftIdGenerator.DontCareId)));
-            result.DatabaseRecord.RevisionsConfigurationUpdated = true;
+                tasks.Add(_server.SendToLeaderAsync(new EditRevisionsConfigurationCommand(databaseRecord.Revisions, _name, RaftIdGenerator.DontCareId)));
+                result.DatabaseRecord.RevisionsConfigurationUpdated = true;
+            }
         }
 
         if (databaseRecord.RevisionsForConflicts != null && databaseRecordItemType.HasFlag(DatabaseRecordItemType.Revisions))
@@ -346,11 +362,19 @@ public sealed class DatabaseRecordActions : IDatabaseRecordActions
 
         if (databaseRecord.Expiration != null && databaseRecordItemType.HasFlag(DatabaseRecordItemType.Expiration))
         {
-            if (_log.IsInfoEnabled)
-                _log.Info("Configuring expiration from smuggler");
+            if (authenticationEnabled && CanAccess(authorizationStatus) == false)
+            {
+                result.AddError("Import of Expiration was skipped due to insufficient permissions on your current certificate.");
+                result.DatabaseRecord.ErroredCount++;
+            }
+            else
+            {
+                if (_log.IsInfoEnabled)
+                    _log.Info("Configuring expiration from smuggler");
 
-            tasks.Add(_server.SendToLeaderAsync(new EditExpirationCommand(databaseRecord.Expiration, _name, RaftIdGenerator.DontCareId)));
-            result.DatabaseRecord.ExpirationConfigurationUpdated = true;
+                tasks.Add(_server.SendToLeaderAsync(new EditExpirationCommand(databaseRecord.Expiration, _name, RaftIdGenerator.DontCareId)));
+                result.DatabaseRecord.ExpirationConfigurationUpdated = true;
+            }
         }
 
         if (databaseRecord.Refresh != null && databaseRecordItemType.HasFlag(DatabaseRecordItemType.Refresh))
@@ -361,7 +385,6 @@ public sealed class DatabaseRecordActions : IDatabaseRecordActions
             tasks.Add(_server.SendToLeaderAsync(new EditRefreshCommand(databaseRecord.Refresh, _name, RaftIdGenerator.DontCareId)));
             result.DatabaseRecord.RefreshConfigurationUpdated = true;
         }
-        
         
         if (databaseRecord.DataArchival != null && databaseRecordItemType.HasFlag(DatabaseRecordItemType.DataArchival))
         {
@@ -374,15 +397,23 @@ public sealed class DatabaseRecordActions : IDatabaseRecordActions
 
         if (databaseRecord.RavenConnectionStrings.Count > 0 && databaseRecordItemType.HasFlag(DatabaseRecordItemType.RavenConnectionStrings))
         {
-            if (_log.IsInfoEnabled)
-                _log.Info("Configuring Raven connection strings configuration from smuggler");
-
-            foreach (var connectionString in databaseRecord.RavenConnectionStrings)
+            if (authenticationEnabled && CanAccess(authorizationStatus) == false)
             {
-                tasks.Add(_server.SendToLeaderAsync(new PutRavenConnectionStringCommand(connectionString.Value, _name, RaftIdGenerator.DontCareId)));
+                result.AddError("Import of Raven Connection Strings was skipped due to insufficient permissions on your current certificate.");
+                result.DatabaseRecord.ErroredCount++;
             }
+            else
+            {
+                if (_log.IsInfoEnabled)
+                    _log.Info("Configuring Raven connection strings configuration from smuggler");
 
-            result.DatabaseRecord.RavenConnectionStringsUpdated = true;
+                foreach (var connectionString in databaseRecord.RavenConnectionStrings)
+                {
+                    tasks.Add(_server.SendToLeaderAsync(new PutRavenConnectionStringCommand(connectionString.Value, _name, RaftIdGenerator.DontCareId)));
+                }
+
+                result.DatabaseRecord.RavenConnectionStringsUpdated = true;
+            }
         }
 
         if (databaseRecord.SqlConnectionStrings.Count > 0 && databaseRecordItemType.HasFlag(DatabaseRecordItemType.SqlConnectionStrings))
@@ -400,11 +431,19 @@ public sealed class DatabaseRecordActions : IDatabaseRecordActions
 
         if (databaseRecord.Client != null && databaseRecordItemType.HasFlag(DatabaseRecordItemType.Client))
         {
-            if (_log.IsInfoEnabled)
-                _log.Info("Configuring client configuration from smuggler");
+            if (authenticationEnabled && CanAccess(authorizationStatus) == false)
+            {
+                result.AddError("Import of Client Configuration was skipped due to insufficient permissions on your current certificate.");
+                result.DatabaseRecord.ErroredCount++;
+            }
+            else
+            {
+                if (_log.IsInfoEnabled)
+                    _log.Info("Configuring client configuration from smuggler");
 
-            tasks.Add(_server.SendToLeaderAsync(new EditDatabaseClientConfigurationCommand(databaseRecord.Client, _name, RaftIdGenerator.DontCareId)));
-            result.DatabaseRecord.ClientConfigurationUpdated = true;
+                tasks.Add(_server.SendToLeaderAsync(new EditDatabaseClientConfigurationCommand(databaseRecord.Client, _name, RaftIdGenerator.DontCareId)));
+                result.DatabaseRecord.ClientConfigurationUpdated = true;
+            }
         }
 
         if (databaseRecord.Integrations?.PostgreSql != null && databaseRecordItemType.HasFlag(DatabaseRecordItemType.PostgreSQLIntegration))
@@ -605,6 +644,13 @@ public sealed class DatabaseRecordActions : IDatabaseRecordActions
         }
 
         tasks.Clear();
+    }
+
+    private static bool CanAccess(AuthorizationStatus authorizationStatus)
+    {
+        return authorizationStatus == AuthorizationStatus.ClusterAdmin ||
+               authorizationStatus == AuthorizationStatus.Operator ||
+               authorizationStatus == AuthorizationStatus.DatabaseAdmin;
     }
 
     public ValueTask DisposeAsync()
